+++ src/core/shared/enums/interiorList.ts
@@ -1,0 +1,8 @@
+const INTERIOR_LIST = {
+    ZancudoBunker: {
+        ipl: 'gr_case10_bunkerclosed',
+        pos: { x: -3058.714, y: 3329.19, z: 12.5844 }
+    }
+};
+
+export default INTERIOR_LIST;